prediction/requirements.txt
Flask==2.0.2
numpy==1.21.2
boto3==1.20.20
py-eureka-client==0.10.0
requests==2.26.0
pytorch-lightning==1.4.4
pytorch-ignite==0.4.7
torchvision==0.10.1
transformers==4.12.5
Pillow=9.0.0
torch=1.10.1
botocore=1.23.31